test(integration): Move test file read and parsing it to Masterdata in a common place
