+++ crates/router/tests/integration/types.rs
@@ -1,9 +1,12 @@
 use actix_web::test::TestRequest;
 use serde_json::value::{Value};
 use serde::{Deserialize, Serialize};
+use std::fs;
+use serde_json;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MasterData{
+// fields derieved from Api responses
   pub merchant_id : Option<String>,
   pub api_key : Option<String>,
   pub api_key_id : Option<String>,
@@ -32,4 +35,30 @@ pub trait RequestBuilder{
   fn verify_success_response(response : &Value, data : &MasterData) -> Self;
   fn verify_failure_response(response : &Value, data : &MasterData) -> Self;
   fn update_master_data(&self,data : &mut MasterData, resp : &Value);
+}
+
+
+fn get_master_data(test_file_path : std::path::PathBuf) -> MasterData{
+  let contents = fs::read_to_string(&test_file_path).expect("Failed to read file");
+  let master_data: MasterData = serde_json::from_str(&contents).expect("Failed to parse JSON");
+  //println!("Initial Master Data : \n {:?}",master_data);
+  return master_data;
+}
+
+pub fn collect_test_data(test_data_dir_path:&str) -> Result<Vec<(String,MasterData)>,String>{
+  if let Ok(test_files) = fs::read_dir(test_data_dir_path) {
+      let mut master_data_list = Vec::new();
+      for test_file in test_files {
+          if let Ok(test_file) = test_file {
+              let test_file_path = test_file.path();
+              if test_file_path.is_file() {
+                let master_data = get_master_data(test_file_path);
+                master_data_list.push((test_file.path().to_string_lossy().into_owned(),master_data));
+              }
+          }
+      }
+    Ok(master_data_list)
+  } else {
+     Err(String::from("Unable to read dir"))
+  }
 }
